seperated db and models
